Added ability to avoid using OS distro packages

+++ manifests/puppetdb/compat.pp
@@ -8,17 +8,19 @@
 # @example
 #   include puppet::puppetdb::compat
 #
-class puppet::puppetdb::compat inherits puppet::params {
-  include puppetdb
+class puppet::puppetdb::compat inherits puppet::puppetdb::globals {
+  $database_ini = "${puppet::puppetdb::globals::confdir}/database.ini"
 
-  $database_ini = "${puppet::params::puppetdb_confdir}/database.ini"
+  if $puppet::globals::compat_mode {
+    include puppetdb
 
-  #  Duplicate configuration entry: [:read-database :subname]
-  exec { "rm -f ${database_ini}":
-    path        => '/usr/bin:/bin',
-    refreshonly => true,
-    onlyif      => "grep -q read-database ${database_ini}",
-    subscribe   => Package[$puppetdb::puppetdb_package],
-    before      => Class['puppetdb::server::database'],
+    #  Duplicate configuration entry: [:read-database :subname]
+    exec { "rm -f ${database_ini}":
+      path        => '/usr/bin:/bin',
+      refreshonly => true,
+      onlyif      => "grep -q read-database ${database_ini}",
+      subscribe   => Package[$puppetdb::puppetdb_package],
+      before      => Class['puppetdb::server::database'],
+    }
   }
 }
